fix: plugins config [#41]

+++ src/lib/api/config.ts
@@ -6,7 +6,8 @@ import {
     BaseDirectory,
     exists,
     mkdir,
-    readTextFile,
+    readFile,
+    writeFile,
     writeTextFile
 } from '@tauri-apps/plugin-fs';
 import { appConfigDir } from '@tauri-apps/api/path';
@@ -20,40 +21,40 @@ export async function doesConfigExist(pluginId: string) {
 export async function getConfig(pluginId: string) {
     const fileName = `${pluginId}.config.json`;
 
-    // if (
-    //     !(await exists(fileName, {
-    //         dir: BaseDirectory.AppConfig
-    //     }))
-    // ) {
-    if (!(await exists(await appConfigDir()))) {
-        await mkdir('', {
-            baseDir: BaseDirectory.AppConfig,
-            recursive: true
+    if (!(await exists(fileName, { baseDir: BaseDirectory.AppConfig }))) {
+        if (!(await exists(await appConfigDir()))) {
+            await mkdir('', {
+                baseDir: BaseDirectory.AppConfig,
+                recursive: true
+            });
+        }
+
+        await writeTextFile(fileName, JSON.stringify({}), {
+            baseDir: BaseDirectory.AppConfig
         });
     }
 
-    await writeTextFile(fileName, JSON.stringify({}), {
-        baseDir: BaseDirectory.AppConfig
-    });
-    // }
-
-    const configTxt = await readTextFile(fileName, {
-        baseDir: BaseDirectory.AppConfig
-    });
+    const configTxt = new TextDecoder().decode(
+        await readFile(fileName, {
+            baseDir: BaseDirectory.AppConfig
+        })
+    );
 
     if (!configTxt.trim()) {
         await writeTextFile(fileName, JSON.stringify({}), {
             baseDir: BaseDirectory.AppConfig
         });
-
-        return {};
     }
 
-    return JSON.parse(configTxt) as unknown;
+    return JSON.parse(configTxt.trim() ? configTxt : '{}') as unknown;
 }
 
 export async function saveConfig(pluginId: string, config: unknown) {
-    await writeTextFile(`${pluginId}.config.json`, JSON.stringify(config), {
-        baseDir: BaseDirectory.AppConfig
-    });
+    await writeFile(
+        `${pluginId}.config.json`,
+        new TextEncoder().encode(JSON.stringify(config)),
+        {
+            baseDir: BaseDirectory.AppConfig
+        }
+    );
 }
